+++ tests/test_backends.py
@@ -283,6 +283,16 @@ def test_backend_methods_2(backend):
     values = tc.backend.convert_to_tensor(np.array([0.0, 4.1, 12.0], dtype=np.float32))
     r = tc.backend.numpy(tc.backend.searchsorted(edges, values))
     np.testing.assert_allclose(r, np.array([1, 2, 4]))
+    p = tc.backend.convert_to_tensor(
+        np.array(
+            [0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.2, 0.4], dtype=np.float32
+        )
+    )
+    r = tc.backend.probability_sample(10000, p, status=np.random.uniform(size=[10000]))
+    _, r = np.unique(r, return_counts=True)
+    np.testing.assert_allclose(
+        r - tc.backend.numpy(p) * 10000.0, np.zeros([10]), atol=100, rtol=1
+    )
 
 
 @pytest.mark.parametrize("backend", [lf("npb"), lf("tfb"), lf("jaxb"), lf("torchb")])